[Refactoring-red] Change useSelector and seperate data into fixture
