+++ src/App.test.jsx
@@ -1,21 +1,23 @@
 import React from 'react';
 
+import { useSelector } from 'react-redux';
+
 import { render } from '@testing-library/react';
 
 import App from './App';
 
+jest.mock('react-redux');
+
+import regions from '../fixtures/regions';
+
 test('App', () => {
-  const regions = [
-    {id: 1, name: '서울'},
-    {id: 2, name: '대전'},
-    {id: 3, name: '대구'},
-    {id: 4, name: '부산'},
-    {id: 5, name: '광주'},
-    {id: 6, name: '강원도'},
-    {id: 7, name: '인천'},
-  ];
+
+  useSelector.mockImplementation((selector) => selector({
+    regions,
+  }));
+
   const { getByText } = render((
-    <App regions={regions} />
+    <App />
   ));
 
   expect(getByText('서울')).not.toBeNull();
