+++ src/RegionsContainer.test.jsx
@@ -6,19 +6,13 @@ import { render } from '@testing-library/react';
 
 import RegionsContainer from './RegionsContainer';
 
+import regions from '../fixtures/regions';
+
 jest.mock('react-redux');
 
 test('RegionsContainer', () => {
   useSelector.mockImplementation((selector) => selector({
-    regions: [
-      {id: 1, name: '서울'},
-      {id: 2, name: '대전'},
-      {id: 3, name: '대구'},
-      {id: 4, name: '부산'},
-      {id: 5, name: '광주'},
-      {id: 6, name: '강원도'},
-      {id: 7, name: '인천'},
-    ],
+    regions,
   }));
   const { getByText } = render((
     <RegionsContainer />
