+++ src/index.jsx
@@ -1,9 +1,12 @@
 import React from 'react';
 import ReactDOM from 'react-dom';
+import { Provider } from 'react-redux';
 
 import App from './App';
 
-const store = createStore(rootReducer)
+import { createStore } from './store';
+
+const store = createStore(rootReducer);
 
 ReactDOM.render(
   <Provider store={store}>
