+++ src/store.js
@@ -1,0 +1,7 @@
+import { createStore } from 'redux';
+
+import reducer from './reducer';
+
+const store = createStore(reducer);
+
+export default store;
